Adjust driver current settings, in particular increase Z current to drive the two steppers
Printer was stuttering, particularly on Z. This was due to the TMC2209 current configuration set for a single stepper, and set conservatively at that.
Increased X, Y, and E similar to what I have for the SKR Mini E3 on my Enders, and increased Z substantially to supply enough to drive both Z motors simultaneously. Value 900 chosen as a guess; will monitor motor temperatures when printing to see if they need adjustment again in the future.