Sort out use of the strain gauge
Replaced the bizarre Creality code with simply re-zeroing the strain gauge at the start of a probe. In the idle() loop, just enabled/disabled z probing when the hot end passes below the optical sensor/above the optical sensor respectively.

Impact is that the idle() loop no longer has a 0.4s pause in it, which was unsatisfactory. Altered configuration to use just one probe per position, and slowed the probing speed down to achieve 0.01mm repeatability (slower would likely yield even tighter variance, but as the printer tolerance is supposedly 0.01mm elsewhere, there's no benefit to be had).